+++ interfaces/index.ts
@@ -3,3 +3,4 @@ export * from "./supplier"
 export * from "./user"
 export * from "./rolePermission"
 export * from "./permission"
+export * from "./role"
